Expose test workflow driver to allow use in integ tests
