+++ src/lib.rs
@@ -11,6 +11,7 @@ pub extern crate assert_matches;
 extern crate tracing;
 
 pub mod protos;
+pub mod test_workflow_driver;
 
 mod activity;
 pub(crate) mod core_tracing;
@@ -36,6 +37,7 @@ pub use core_tracing::tracing_init;
 pub use pollers::{PollTaskRequest, ServerGateway, ServerGatewayApis, ServerGatewayOptions};
 pub use protosext::IntoCompletion;
 pub use url::Url;
+pub use workflow::CommandID;
 
 use crate::{
     activity::{ActivityHeartbeatManager, ActivityHeartbeatManagerHandle, InflightActivityDetails},
