+++ src/machines/test_help/mod.rs
@@ -4,7 +4,6 @@ mod async_workflow_driver;
 mod history_builder;
 mod transition_coverage;
 
-pub(super) use async_workflow_driver::{CommandSender, TestWorkflowDriver};
 pub(crate) use history_builder::TestHistoryBuilder;
 pub(crate) use transition_coverage::add_coverage;
